+++ table/table_builder.h
@@ -37,9 +37,9 @@ struct TableReaderOptions {
       const std::shared_ptr<const SliceTransform>& _prefix_extractor,
       const EnvOptions& _env_options,
       const InternalKeyComparator& _internal_comparator,
-      bool _skip_filters = false, bool _immortal = false,
-      bool _force_direct_prefetch = false, int _level = -1,
-      BlockCacheTracer* const _block_cache_tracer = nullptr,
+      uint8_t _block_protection_bytes_per_key, bool _skip_filters = false,
+      bool _immortal = false, bool _force_direct_prefetch = false,
+      int _level = -1, BlockCacheTracer* const _block_cache_tracer = nullptr,
       size_t _max_file_size_for_l0_meta_pin = 0,
       const std::string& _cur_db_session_id = "", uint64_t _cur_file_num = 0,
       UniqueId64x2 _unique_id = {}, SequenceNumber _largest_seqno = 0)
@@ -56,7 +56,8 @@ struct TableReaderOptions {
         max_file_size_for_l0_meta_pin(_max_file_size_for_l0_meta_pin),
         cur_db_session_id(_cur_db_session_id),
         cur_file_num(_cur_file_num),
-        unique_id(_unique_id) {}
+        unique_id(_unique_id),
+        block_protection_bytes_per_key(_block_protection_bytes_per_key) {}
 
   const ImmutableOptions& ioptions;
   const std::shared_ptr<const SliceTransform>& prefix_extractor;
@@ -86,6 +87,8 @@ struct TableReaderOptions {
 
   // Known unique_id or {}, kNullUniqueId64x2 means unknown
   UniqueId64x2 unique_id;
+
+  uint8_t block_protection_bytes_per_key;
 };
 
 struct TableBuilderOptions {